죽겠다.. 오늘이 마지막인가..

+++ WebProject5/src/selfGuide/controller/SelfGuideRemoveServlet.java
@@ -2,12 +2,15 @@ package selfGuide.controller;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import selfGuide.model.service.SelfGuidePhotoService;
 import selfGuide.model.service.SelfGuideService;
 
 /**
@@ -34,13 +37,30 @@ public class SelfGuideRemoveServlet extends HttpServlet {
 		String saveDirectory = root + "upload/photo";
 		String fileName = request.getParameter("fileName");
 		int selfNo = Integer.parseInt(request.getParameter("selfNo"));
+		
 		String filePath = saveDirectory+"/"+fileName;
 		System.out.println(filePath);
 		
 		File file = new File(filePath);
 		
 		int result = new SelfGuideService().guideRemove(selfNo);
-		if(result > 0) {
+		
+		ArrayList<String> removeValueList = new SelfGuidePhotoService().guidePhotoRemoveValue(selfNo);
+		if(!removeValueList.isEmpty()) {
+			for(String removeValue : removeValueList) {
+				System.out.println("성공?");
+				System.out.println(removeValue);
+				File file2 = new File(removeValue);
+				file2.delete();
+			}
+		} else {
+			System.out.println("풒킥풒킥");
+		}
+		
+		
+		int result2 = new SelfGuidePhotoService().guidePhotoRemove(selfNo);
+		
+		if(result > 0 && result2 >0) {
 			file.delete();
 			response.sendRedirect("/views/selfGuide/selfGuideMain.jsp");
 		} else {
